Code/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.5
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
oauthlib==3.2.2
python-dotenv==0.21.1
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.3
Werkzeug==2.2.3
